[test/create-for-stream] Add the missing cairo_surface_finish().
Moments after pushing the new test case did I realise the issue...
We do not attempt to write out the surface to the user stream until
we perform the cairo_surface_destroy() by which point we have lost
the ability to interrogate the error status. We can avoid this by
explicitly calling cairo_surface_finish() and then checking the
error status - and we see that the error is indeed reported
correctly.

No bug. Nothing to see here. Please move along. (Apart from the
request for the status to be return from cairo_surface_destroy!)