[cairo-type1-subset] Propagate error status (correctly!)
Adrian Johnson pointed out a couple of mistakes in my previous attempt,
535e7c161b907292eac098b6b4305373558948df, to ensure propagation of errors
throughout the generation of the type1 subset.

This time the status member of the cairo_type1_font_subset_t is removed
in favour of simply return the error status from each function. This
completely avoids the issue of whether we overwrite a pre-existing error
status and confustion of status returns and the status member. The
removal of the status from the structure is possible due to its
short-lived nature - it is not exposed outside of the
_cairo_type1_subset_init() function, and is not shared by any other
piece of code.