Use correct surface resolution when creating paginated fallback image
If a surface resolution is set via _cairo_surface_set_resolution, it
needs to be used when the fallback image is created.  The user sets
the fallback resolution on the paginated surface, so we need
to divide that by the resolution of the actual (target) surface, not
the paginated surface.