Implement a win32 tmpfile() function
Bug report and proposed patch at:
http://lists.cairographics.org/archives/cairo/2007-December/012529.html

On Windows the tmpfile() function creates the file in the root
directory. This will fail if the user does not have write access to the
root directory.

Implement _cairo_win32_tmpfile() that is #defined as tmpfile() on
Windows. This function uses GetTempPath() and GetTempFileName() to
create the temporary file. CreateFile() is used to open the file so
the DELETE_ON_CLOSE flag can be set.