Fix PDF extend-reflect test failure
Previously, when emitting image patterns the PDF backend used
_cairo_pattern_acquire_surface to get the image. For reflected images
this would return an image containing four images in a reflect
pattern.  When drawn in a PDF pattern (which only does repeating
patterns) this would create the reflected pattern in PDF.

For some reason _cairo_pattern_acquire_surface is no longer returning
a reflected image pattern.

This is fixed by only using _cairo_surface_acquire_source_image to get
the image and using the same code as is used for reflected
meta-surface patterns to created a reflected pattern by drawing four
transformed copies of the image inside the PDF patten.

This is the better way to implement reflected images as we are no
longer embedding an image four times larger than the original.