svg: Add support for flattening SOURCE painting when there's nothing beneath it
This avoids unnecessary rasterization in many cases when using
cairo_surface_create_similar with an SVG surface. Because of that
it eliminates test-suite failures for the -similar cases where we
have svg-specific reference images. Namely:

	font-matrix-translation, ft-text-vertical-layout-type1,
	ft-text-vertical-layout-type3, mask, meta-surface-pattern,
	paint-source-alpha, paint-with-alpha, rotate-image-surface-paint,
	scale-source-surface-paint, source-clip-scale, text-pattern,
	text-rotate

In all of these cases the test suite was kindly noticing that we
weren't getting the same 'native' SVG output that was desired.