Track minor change in pixman filtering sample location
This change in reference image was triggered by the following commit
in the pixman library:

	commit f2d8a5caa63596739b4dc47c00daaaafd9f522e2
	Author: Søren Sandmann <sandmann@redhat.com>
	Date:   Mon Mar 10 23:41:52 2008 -0400

	Add rounding epsilon for NEAREST filter after transformation, not before.

This hasn't appeared in a pixman release yet. So for now, getting
correct results is a matter of "use latest pixman". Soon we'll
have a new pixman release and we'll make cairo depend on that.